Need an extra check for torrent's artist before deleting it in this transitional stage

+++ app/models/torrent.rb
@@ -52,7 +52,7 @@ class Torrent < ActiveRecord::Base
   end
 
   def destroy_with_artist
-    if self.artist_id.present? and self.artist.torrents.count == 1
+    if self.artist_id.present? and self.artist.present? and self.artist.torrents.count == 1
       self.artist.destroy
     end
